requirements.txt
streamlit
word2number
Pillow==6.2.2
spacy

